[refactor] Change dual pane mode implementation
