--- app/src/main/java/be/hogent/faith/faith/library/EventDetailActivity.kt
@@ -1,66 +1,0 @@
-package be.hogent.faith.faith.library
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.MenuItem
-import androidx.appcompat.app.AppCompatActivity
-import be.hogent.faith.R
-
-/**
- * An activity representing a single Event detail screen. This
- * activity is only used on narrow width devices. On tablet-size devices,
- * item details are presented side-by-side with a list of items
- * in a [EventListActivity].
- */
-class EventDetailActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_event_detail)
-
-
-        // Show the Up button in the action bar.
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-
-        // savedInstanceState is non-null when there is fragment state
-        // saved from previous configurations of this activity
-        // (e.g. when rotating the screen from portrait to landscape).
-        // In this case, the fragment will automatically be re-added
-        // to its container so we don't need to manually add it.
-        // For more information, see the Fragments API guide at:
-        //
-        // http://developer.android.com/guide/components/fragments.html
-        //
-        if (savedInstanceState == null) {
-            // Create the detail fragment and add it to the activity
-            // using a fragment transaction.
-            val fragment = EventDetailFragment().apply {
-                arguments = Bundle().apply {
-                    putString(
-                        EventDetailFragment.ARG_ITEM_ID,
-                        intent.getStringExtra(EventDetailFragment.ARG_ITEM_ID)
-                    )
-                }
-            }
-
-            supportFragmentManager.beginTransaction()
-                .add(R.id.event_detail_container, fragment)
-                .commit()
-        }
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem) =
-        when (item.itemId) {
-            android.R.id.home -> {
-                // This ID represents the Home or Up button. In the case of this
-                // activity, the Up button is shown. For
-                // more details, see the Navigation pattern on Android Design:
-                //
-                // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-
-                navigateUpTo(Intent(this, EventListActivity::class.java))
-                true
-            }
-            else -> super.onOptionsItemSelected(item)
-        }
-}
