+++ app/src/main/java/be/hogent/faith/faith/library/EventDetailFragment.kt
@@ -6,8 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import be.hogent.faith.R
-import be.hogent.faith.faith.library.dummy.DummyContent
-import kotlinx.android.synthetic.main.event_detail.view.event_detail
+
 
 /**
  * A fragment representing a single Event detail screen.
@@ -17,43 +16,16 @@ import kotlinx.android.synthetic.main.event_detail.view.event_detail
  */
 class EventDetailFragment : Fragment() {
 
-    /**
-     * The dummy content this fragment is presenting.
-     */
-    private var item: DummyContent.DummyItem? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        arguments?.let {
-            if (it.containsKey(ARG_ITEM_ID)) {
-                // Load the dummy content specified by the fragment
-                // arguments. In a real-world scenario, use a Loader
-                // to load content from a content provider.
-                item = DummyContent.ITEM_MAP[it.getString(ARG_ITEM_ID)]
-            }
-        }
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val rootView = inflater.inflate(R.layout.event_detail, container, false)
+        val rootView = inflater.inflate(R.layout.fragment_eventdetails, container, false)
 
-        // Show the dummy content as text in a TextView.
-        item?.let {
-            rootView.event_detail.text = it.details
-        }
 
         return rootView
     }
 
-    companion object {
-        /**
-         * The fragment argument representing the item ID that this fragment
-         * represents.
-         */
-        const val ARG_ITEM_ID = "item_id"
-    }
+
 }
