+++ app/src/main/res/layout/activity_eventlist.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/layout_events_details"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:fitsSystemWindows="true"
-    tools:context=".faith.library.EventDetailActivity"
-    tools:ignore="MergeRootFrame">
+    tools:context=".faith.library.EventListActivity">
 
 
-    <androidx.recyclerview.widget.RecyclerView
+    <fragment
+        android:id="@+id/fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:layout_editor_absoluteX="171dp"
-        tools:layout_editor_absoluteY="215dp" />
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph_events_single" />
 </androidx.constraintlayout.widget.ConstraintLayout>
