+++ app/src/main/res/navigation/nav_graph_events_dual.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/nav_graph_events_dual"
+    app:startDestination="@id/eventDetailFragment">
+    <fragment
+        android:id="@+id/eventDetailFragment"
+        android:name="be.hogent.faith.faith.library.EventDetailFragment"
+        android:label="EventDetailFragment" />
+</navigation>
